[AST] split AST.cpp
